`matchadd` on a per-window basis
Vim's `:h matchadd()` states matches are added on a per-window basis.
Therefore we should not add matches if new buffer is opened/created
because it might add duplicate matches in the same vim window.

The final solution to this seems to be to keep track whether a match was
added to a window or not.